+++ lilo/src/main/java/com/amrdeveloper/lilo/object/LiloComplex.kt
@@ -2,6 +2,6 @@ package com.amrdeveloper.lilo.`object`
 
 import com.amrdeveloper.lilo.type.liloComplexType
 
-data class LiloComplex(val real: Float, val imag: Float) : LiloObject(liloComplexType) {
+data class LiloComplex(val real: Double, val imag: Double) : LiloObject(liloComplexType) {
     override fun toString() = "($real+${imag}j)"
 }
